### 2022-04-04 #4297 @editorialbot
[PRE REVIEW]: GHEtool: An open-source tool for borefield sizing in Python
**Submitting author:** @wouterpeer (<a href="http://orcid.org/0000-0002-2311-5981">Wouter Peere</a>)
**Repository:** https://github.com/wouterpeere/GHEtool
**Branch with paper.md** (empty if default branch): main
**Version:** v2.0.0
