+++ Paper.bib
@@ -1,0 +1,65 @@
+@INPROCEEDINGS{GLHEPRO,
+    author = {Jeffrey D. Spitler},
+    title = {{GLHEPRO}-- A design tool for commercial building ground loop heat exchangers},
+    booktitle = {Proceedings of the fourth international heat pumps in cold climates conference},
+    year = {2000},
+    note={Aylmer, Québec. August 17-18, 2000.}},
+
+@article{Bernier,
+	title = "A review of vertical ground heat exchanger sizing tools including an inter-model comparison",
+	journal = "Renewable and Sustainable Energy Reviews",
+	volume = "110",
+	pages = "247 - 265",
+	year = "2019",
+	issn = "1364-0321",
+	doi = "https://doi.org/10.1016/j.rser.2019.04.045",
+	url = "http://www.sciencedirect.com/science/article/pii/S1364032119302576",
+	author = "Mohammadamin Ahmadfard and Michel Bernier",
+	keywords = "Geothermal energy, Sizing tools, Vertical ground heat exchangers, Inter-model comparison, Test cases"},
+
+@inproceedings{gfunctions,
+  title={Thermal analysis of heat extraction boreholes},
+  author={Per Eskilson},
+  year={1987},
+  booktitle    = {PhD thesis, Dep. of Mathematical Physics, University of Lund, Sweden}},
+
+@inproceedings{pygfunction,
+  title={pygfunction: an open-source toolbox for the evaluation of thermal response factors for geothermal borehole fields},
+  author={Massimo Cimmino},
+  year={2018},
+  booktitle    = {Proceedings of eSim 2018, the 10th conference of IBPSA-Canada},
+  note = {Montréal, QC, Canada, May 9-10}},
+
+@article{EED,
+  title={Earth energy designer},
+  author={Hellstr{\"o}m, G{\"o}ran and Sanner, Burkhard},
+  journal={User’s Manual, version},
+  volume={2},
+  number={20},
+  pages={0},
+  year={2000}},
+
+@INPROCEEDINGS{Peere,
+    author = {Wouter {Peere} and Damien {Picard} and Iago {Cupeiro~Figueroa} and Wim {Boydens} and Lieve {Helsen}},
+    title = {Validated combined first and last year borefield sizing methodology},
+    booktitle = {International Building Simulation  Conference 2021},
+    year = {2021},
+    note = {Brugge, Belgium, 1-3 September 2021}},
+
+@article{IEA,
+  title={GLOBAL STATUS REPORT: Towards a zero-emission, efficient, and resilient buildings and construction sector},
+  author={Abergel, Thibaut and Dean, Brian and Dulac, John and Hamilton, Ian and Wheeler, T},
+  journal={Global Alliance for Buildings and Construction. Available online: https://www. worldgbc. org/sites/default/files/2018\% 20GlobalABC\% 20Global\% 20Status\% 20Report (accessed on 28 February 2022)},
+  year={2018}},
+
+@misc{geoSIM,
+	author = {{geoSIM}},
+	title = {{geoSIM – your freeware tool for simulating and sizing of borefields with geoKOAX}},
+	note = {Available online: https://geokoax.com/en/our-products/simulation/ (accessed on 10 February 2022)},
+	url = {https://geokoax.com/en/our-products/simulation/}},
+
+@misc{igshpa,
+	author = {{Igshpa}},
+	title = {{Software – Tools to design and support your Ground Source Heat Pump Installation}},
+	note = {Available online: https://igshpa.org/software/ (accessed on 10 February 2022)},
+	url = {https://igshpa.org/software/}}
