+++ Paper.md
@@ -1,0 +1,89 @@
+---
+title: 'GHEtool: An open-source tool for borefield sizing in Python'
+tags:
+  - geothermal
+  - energy
+  - borefields
+  - sizing
+authors:
+  - name: Wouter Peere^[corresponding author]
+    orcid: 0000-0002-2311-5981
+    affiliation: "1, 2" 
+  - name: Tobias Blanke
+    orcid: 0000-0003-1529-5529
+    affiliation: 3 
+affiliations:
+ - name: Department of Mechanical Engineering, University of Leuven (KU Leuven), Leuven, Belgium
+   index: 1
+ - name: boydens engineering - part of Sweco, Dilbeek, Belgium
+   index: 2
+ - name: Solar-Institute Jülich, FH Aachen, Aachen, Germany
+   index: 3
+date: 01 April 2022
+bibliography: Paper.bib
+
+---
+
+# Summary
+
+GHEtool is a python package that contains all the functionalities needed to deal with borefield design. It is developed for both researchers and practitioners.
+The core of this package is the automated sizing of borefield under different conditions.
+The sizing of a borefield is typically slow due to the high complexity of the mathematical background. Because this tool has a lot of precalculated data, GHEtool can size a borefield in the order of tenths of milliseconds.
+This sizing typically takes the order of minutes. Therefore, this tool is suited for being implemented in typical workflows where iterations are required.
+
+GHEtool also comes with a graphical user interface (GUI). This GUI is prebuilt as an exe-file because this provides access to all the functionalities without coding.
+A setup to install the GUI at the user-defined place is also implemented and available at: [https://www.mech.kuleuven.be/en/tme/research/thermal_systems/ghetool](https://www.mech.kuleuven.be/en/tme/research/thermal_systems/ghetool).
+
+
+# Statement of need
+
+The building sector uses 36 % of global energy and is responsible for 39 % of energy-related CO2 emissions when upstream power generation is included, with more than half of this energy used by space heating, space cooling and water heating [@IEA].
+This sector should decarbonize to reach the climate goals in 2050. One of the promising ways to achieve this goal is by using shallow geothermal energy, thereby storing heat/cold in borefields, enabling seasonal thermal energy storage.
+Therefore, much interest exists from academia and practitioners in this field.
+
+One of the main challenges in this domain is to size a borefield, which is critical since borefields have a very high investment cost, so we want them to be as small as possible to increase their (economic) feasibility.
+In all research related to optimising borefield loads, studying sensitivity on borefield configurations, and integrating seasonal thermal energy storage in fifth-generation district heating etc. borefield sizing plays a central role.
+
+GHEtool is a python package with a graphical counterpart centred around borefield sizing and the evaluation of temperature evolution. 
+
+The main advantage of GHEtool is that the ground response functions (gfunctions [@gfunctions]), which are needed to size a borefield, are already precalculated for more than 140 000 borefield – ground combinations.
+Due to this precalculation, one can save from a couple of seconds to several minutes to calculate the gfunction for even one borefield. On the other hand, the tool provides the option to provide your gfunction data for your own custom borefield configuration. 
+
+GHEtool aims to be used both by academic researchers in thermal systems integration and as a tool for educational purposes. It offers all the code and functions needed to size a borefield.
+Also, it evaluates the mean fluid and borehole wall temperature evolution. Therefore, GHEtool is beneficial for research activities relating to fifth-generation district heating, seasonal thermal energy storage and sensitivity analysis on borefield parameters.
+
+The tool comes with a graphical user interface. This interface is particularly beneficial for educational purposes, where students (and trained engineers) can play around to see how specific decisions influence the behaviour of the borefield.
[…]
