### 2022-04-07 #4297 @crvernon
:wave: @wouterpeere - I'll be your topic editor for this submission.

The first step in this pre-review will be to find at least 2 reviewers and get the review started. If you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). In addition, this [list](https://docs.google.com/spreadsheets/d/1PAPRJ63yq9aPC1COLjaQp8mHmEq3rZUzwUYxTulyu78/edit#gid=856801822) of people have already agreed to review for JOSS and may be suitable for this submission (please start at the bottom of the list). Thanks!
