### 2022-05-13 #4297 @crvernon
👋 - @raoulcollenteur Would you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
