### 2022-05-19 #4297 @crvernon
👋 - @nmstreethran , @mefuller , or @fwitte Would either of you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
