+++ paper.bib
@@ -12,7 +12,7 @@
 	pages = "247 - 265",
 	year = "2019",
 	issn = "1364-0321",
-	doi = "https://doi.org/10.1016/j.rser.2019.04.045",
+	doi = "10.1016/j.rser.2019.04.045",
 	url = "http://www.sciencedirect.com/science/article/pii/S1364032119302576",
 	author = "Mohammadamin Ahmadfard and Michel Bernier",
 	keywords = "Geothermal energy, Sizing tools, Vertical ground heat exchangers, Inter-model comparison, Test cases"},
