### 2022-06-08 wouterpeere/GHEtool#7 @nmstreethran
[JOSS Review] Functionality + documentation
Hi @wouterpeere, I'm one of the reviewers assigned to your submission to JOSS (https://github.com/openjournals/joss-reviews/issues/4406). I'm opening this issue as part of my review; it concerns the functionality and documentation of GHEtool. The headings below correspond to items in my review checklist. Please let me know if anything is unclear. Thanks.

## Installation / installation instructions

> Does installation proceed as outlined in the documentation?
> Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.

The installation instruction is currently limited to the quick start using pip, and it states that "Developers can clone this repository". I highly recommend including some additional steps here (such as cloning the repository and using a [virtual environment](https://docs.python.org/3.9/library/venv.html)) to help users new to Python.

Could you also add a link or badge to the [PyPI package](https://pypi.org/project/GHEtool/) in the README?

## Functionality / functionality documentation

> Have the functional claims of the software been confirmed?
> Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

Functionality documentation seems to be missing. Sufficient API documentation is required, and the core API should be documented at the very least. You can add this to the README.

## Automated tests

> Are there automated tests or manual steps described so that the functionality of the software can be verified?

I can't seem to find any tests. Have you implemented any tests and what are the steps to run them? I see that you have some scripts in the `GHEtool/Validation` folder. The testing instructions should be clearly stated so that users can determine whether the functions work after installing the software.

## Example usage

> Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

You have included examples in the `GHEtool/Examples` folder. Can you link each of these scripts to the list of functionalities of the software in the README, and add simple instructions on how users can run these scripts?
