### 2022-06-09 b03423f
Badge PyPi and information about venv
