### 2022-06-09 wouterpeere/GHEtool#7 @wouterpeere
Hi @nmstreethran, thank you for your comments.

W.r.t. the installation: I wrote a small paragraph under the installation section with a link to the official python documentation and an article on venv that I found very clear. I also included a PyPi-badge, thank you for this tip!

W.r.t. the functionality: I wrote most of the API documentation in the example documents, but perhaps that is not the best location. Would it be better if I took some code snippets and placed them directly on the README? I can then also directly refer to the different examples.

W.r.t. the automated test. The scripts in the validation folder are more about the 'correctness' of the results from the package than about whether or not the package is installed correctly. I will make a test for that and explain (perhaps also under the installation section) how one can run these tests.

W.r.t. the examples. Yes, I will refer to them in the README (like mentioned hereabove).
