### 2022-06-09 cc0af0b
Bufix w.r.t. zero peak loads
