+++ GHEtool/test/test_GHEtool.py
@@ -1,0 +1,103 @@
+# test if model can be imported
+from GHEtool import *
+import pytest
+
+def test_grounddata():
+    data = GroundData(110, 6, 3, 10, 0.2, 10, 12)
+    assert data.H == 110
+    assert data.B == 6
+    assert data.k_s == 3
+    assert data.Tg == 10
+    assert data.Rb == 0.2
+    assert data.N_1 == 10
+    assert data.N_2 == 12
+
+def test_fluiddata():
+    data = FluidData(0.2, 0.568, 998, 4180, 1e-3)
+    assert data.mfr == 0.2
+    assert data.k_f == 0.568
+    assert data.rho == 998
+    assert data.Cp == 4180
+    assert data.mu == 1e-3
+
+def test_pipedata():
+    data = PipeData(1, 0.015, 0.02, 0.4, 0.05, 0.075, 2)
+    assert data.k_g == 1
+    assert data.r_in == 0.015
+    assert data.r_out == 0.02
+    assert data.k_p == 0.4
+    assert data.D_s == 0.05
+    assert data.numberOfPipes == 2
+    assert data.D == 4
+    assert data.r_b == 0.075
+
+data = GroundData(110, 6, 3, 10, 0.2, 10, 12)
+
+# Montly loading values
+peakCooling = [0., 0, 34., 69., 133., 187., 213., 240., 160., 37., 0., 0.]              # Peak cooling in kW
+peakHeating = [160., 142, 102., 55., 0., 0., 0., 0., 40.4, 85., 119., 136.]             # Peak heating in kW
+
+# annual heating and cooling load
+annualHeatingLoad = 300*10**3 # kWh
+annualCoolingLoad = 160*10**3 # kWh
+
+# percentage of annual load per month (15.5% for January ...)
+montlyLoadHeatingPercentage = [0.155, 0.148, 0.125, .099, .064, 0., 0., 0., 0.061, 0.087, 0.117, 0.144]
+montlyLoadCoolingPercentage = [0.025, 0.05, 0.05, .05, .075, .1, .2, .2, .1, .075, .05, .025]
+
+# resulting load per month
+monthlyLoadHeating = list(map(lambda x: x * annualHeatingLoad, montlyLoadHeatingPercentage))   # kWh
+monthlyLoadCooling = list(map(lambda x: x * annualCoolingLoad, montlyLoadCoolingPercentage))   # kWh
+
+def test_borefield():
+    borefield = Borefield(simulationPeriod=20,
+                          peakHeating=peakHeating,
+                          peakCooling=peakCooling,
+                          baseloadHeating=monthlyLoadHeating,
+                          baseloadCooling=monthlyLoadCooling)
+
+    borefield.setGroundParameters(data)
+
+    # set temperature boundaries
+    borefield.setMaxGroundTemperature(16)  # maximum temperature
+    borefield.setMinGroundTemperature(0)  # minimum temperature
+
+    assert borefield.simulationPeriod == 20
+    assert borefield.Tf_C == 0
+    assert borefield.Tf_H == 16
+    assert borefield.peakHeating == [160., 142, 102., 55., 26.301369863013697, 0., 0., 0., 40.4, 85., 119., 136.]
+
+@pytest.fixture
+def borefield():
+    borefield = Borefield(simulationPeriod=20,
+                          peakHeating=peakHeating,
+                          peakCooling=peakCooling,
+                          baseloadHeating=monthlyLoadHeating,
+                          baseloadCooling=monthlyLoadCooling)
+
+    borefield.setGroundParameters(data)
+
+    # set temperature boundaries
+    borefield.setMaxGroundTemperature(16)  # maximum temperature
+    borefield.setMinGroundTemperature(0)  # minimum temperature
[…]
