+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = GHEtool
-version = 2.0.0
+version = 2.0.1
 author = Wouter Peere
 author_email = wouter.peere@kuleuven.be
 description = Python package for borefield sizing
