### 2022-06-12 f90318f
Update README.md
Added getting started under quick start
