### 2022-06-12 43022a0
New example for size borefield by length and width Included new, optimised, function for that
