+++ README.md
@@ -126,7 +126,7 @@ GHEtool offers functionalities of value to all different disciplines working wit
 | Calculating the temperature evolution of the ground for a given building load and borefield configuration | [Main_Functionalities.py](https://github.com/wouterpeere/GHEtool/blob/main/GHEtool/Examples/Main_Functionalities.py) |
 | Using dynamically calculated borehole thermal resistance (this is directly based on the code of pygfunction) | [Sizing_With_Rb_Calculation.py](https://github.com/wouterpeere/GHEtool/blob/main/GHEtool/Validation/Sizing_With_Rb_Calculation.py) |
 | Optimising the load profile for a given heating and cooling load | [Optimise_Load_Profile.py](https://github.com/wouterpeere/GHEtool/blob/main/GHEtool/Examples/Optimise_Load_Profile.py)|
-| Finding the optimal rectangular borefield configuration for a given heating and cooling load | |
+| Finding the optimal rectangular borefield configuration for a given heating and cooling load | [Size_Borefield_By_Length_And_Width](https://github.com/wouterpeere/GHEtool/blob/main/GHEtool/Examples/Size_Borefield_By_Length_And_Width.py) |
 | Importing heating and cooling loads from .csv and .xlsx files | [Import_Data.py](https://github.com/wouterpeere/GHEtool/blob/main/GHEtool/Examples/Import_Data.py) |
 | Using your custom borefield configuration | [Custom_Borefield_Configuration.py](https://github.com/wouterpeere/GHEtool/blob/main/GHEtool/Examples/Custom_Borefield_Configuration.py) |
 
