### 2022-06-12 wouterpeere/GHEtool#7 @wouterpeere
Hi @nmstreethran

I implemented the suggested changes!
