### 2022-06-15 wouterpeere/GHEtool#7 @nmstreethran
Hi @wouterpeere, thank you for implementing the changes.

I don't see the requirements defined anywhere in the package (other than the list in the README). When I follow the installation instructions and install GHEtool with pip or build from source, the requirements are not installed automatically. Could you define them under `install_requires` as shown [here](https://setuptools.pypa.io/en/latest/setuptools.html#setup-cfg-only-projects) and [here](https://setuptools.pypa.io/en/latest/userguide/declarative_config.html)? Pytest should also be included in the list of development requirements (i.e. under `extras_require`).

I also noticed that the PyPI version of GHEtool is 2.0.2, but `setup.cfg` says 2.0.1.

Could you clarify how the requirement for Tkinter is satisfied?
