### 2022-06-15 wouterpeere/GHEtool#7 @nmstreethran
@wouterpeere Can I also suggest you to use relative links to repository files in the README? Currently, you're using the full path, e.g. `https://github.com/wouterpeere/GHEtool/blob/main/GHEtool/Examples/Custom_Borefield_Configuration.py`; this always points to the `main` branch. Use relative links, e.g. `GHEtool/Examples/Custom_Borefield_Configuration.py`.
