### 2022-06-15 wouterpeere/GHEtool#7 @nmstreethran
@wouterpeere Additionally, I had an issue with installing pygfunction as one of its dependencies, CoolProp, fails to install. Looks like this is a problem when installing the current version of CoolProp (v6.4.1) with Python >= 3.9. (I'm on Arch Linux.) Installing from source/testpypi resolved this issue:

```sh
pip install -i https://test.pypi.org/simple/ CoolProp==6.4.2.dev0
pip install GHEtool numpy scipy matplotlib pygfunction openpyxl pandas
```

It may be worth pointing this out in your README (as you say the code is tested with Python 3.9; the other option is to downgrade to 3.8).

See:

- <https://github.com/CoolProp/CoolProp/issues/1992>
- <https://github.com/CoolProp/CoolProp/issues/2119>
