### 2022-06-16 9d9a066
Remove Tkinter
