### 2022-06-16 wouterpeere/GHEtool#7 @nmstreethran
Update re: Tkinter @wouterpeere

I managed to get it to work by first installing Tk on my system. The instructions are platform-specific so I think you can add this to the README with a link to the TkDocs tutorial for installing Tk: https://tkdocs.com/tutorial/install.html
