### 2022-06-16 wouterpeere/GHEtool#7 @wouterpeere
Hi @nmstreethran, thanks for all the comments!

> @wouterpeere Can I also suggest you to use relative links to repository files in the README? Currently, you're using the full path, e.g. `https://github.com/wouterpeere/GHEtool/blob/main/GHEtool/Examples/Custom_Borefield_Configuration.py`; this always points to the `main` branch. Use relative links, e.g. `GHEtool/Examples/Custom_Borefield_Configuration.py`.

This has been changed!

> I don't see the requirements defined anywhere in the package (other than the list in the README). When I follow the installation instructions and install GHEtool with pip or build from source, the requirements are not installed automatically. Could you define them under `install_requires` as shown [here](https://setuptools.pypa.io/en/latest/setuptools.html#setup-cfg-only-projects) and [here](https://setuptools.pypa.io/en/latest/userguide/declarative_config.html)? Pytest should also be included in the list of development requirements (i.e. under `extras_require`).

This has been implemented and uploaded to test PyPi. You can install/test it using the command:
```
pip install --extra-index-url https://test.pypi.org/simple/ GHEtool==2.0.3rc1
```
> I also noticed that the PyPI version of GHEtool is 2.0.2, but `setup.cfg` says 2.0.1.

This has been changed to 2.0.3rc1. After the whole JOSS process is finished, I will set it to 2.0.3, since I realised that otherwise, there will still be a lot of versions to go.

> Could you clarify how the requirement for Tkinter is satisfied?

This was necessary for an old functionality before the GUI existed. I now deleted it, so it is no longer needed.

> Update re: Tkinter @wouterpeere
>
> I managed to get it to work by first installing Tk on my system. The instructions are platform-specific so I think you can add this to the README with a link to the TkDocs tutorial for installing Tk: https://tkdocs.com/tutorial/install.html

This is hence also solved. Sorry for the inconvenience.
