### 2022-06-16 c01a80f
Included dependencies in setup.cfg Changed README
