+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = GHEtool
-version = 2.0.2
+version = 2.0.3rc1
 author = Wouter Peere
 author_email = wouter.peere@kuleuven.be
 description = Python package for borefield sizing
@@ -16,3 +16,14 @@ classifiers =
 packages = find:
 python_requires = >=3.7
 include_package_data = True
+install_requires =
+    numpy>=1.20.2
+	scipy>=1.6.2
+	matplotlib>=3.4.1
+	pygfunction==1.1.2
+	openpyxl>=3.0.7
+	pandas>=1.2.4
+	pytest>=7.1.2
+
+[options.extras_require]
+GUI = pyqt5>=5.10
