### 2022-06-16 wouterpeere/GHEtool#7 @wouterpeere
> @wouterpeere Additionally, I had an issue with installing pygfunction as one of its dependencies, CoolProp, fails to install. Looks like this is a problem when installing the current version of CoolProp (v6.4.1) with Python >= 3.9. (I'm on Arch Linux.) Installing from source/testpypi resolved this issue:
>
> ```shell
> pip install -i https://test.pypi.org/simple/ CoolProp==6.4.2.dev0
> pip install GHEtool numpy scipy matplotlib pygfunction openpyxl pandas
> ```
>
> It may be worth pointing this out in your README (as you say the code is tested with Python 3.9; the other option is to downgrade to 3.8).
>
> See:
>
> * [Installation errors with Python 3.9 CoolProp/CoolProp#1992](https://github.com/CoolProp/CoolProp/issues/1992)
> * [Python bindings: Call for help from the community CoolProp/CoolProp#2119](https://github.com/CoolProp/CoolProp/issues/2119)

I am aware of this problem. If you work with pygfunction==1.1.2, it works in python>=3.9. In the setup.cfg, I set the version of pygfunction to 1.1.2 in order to avoid any problems. I will add a small paragraph related to the problem you mentioned. Thanks for pointing it out!
