### 2022-06-16 4de0120
Note about CoolProp
