### 2022-06-16 wouterpeere/GHEtool#7 @nmstreethran
Hi @wouterpeere, thanks a lot for all the changes!

> This has been changed to 2.0.3rc1. After the whole JOSS process is finished, I will set it to 2.0.3, since I realised that otherwise, there will still be a lot of versions to go.

Good call!

I have installed the package from testPyPI and ran the tests. Everything looks OK, all 9 tests pass but had 22 warnings. This seems to be from Pygfunction v1.1.2 using the deprecated `np.asscalar(a)` and has been fixed in v2.1.0 according to their [changelog](https://github.com/MassimoCimmino/pygfunction/blob/master/CHANGELOG.md#version-210-2021-11-12). I think this is fine though as it's not from your package. Just wanted to point this out.

I haven't finished going through the examples, but I'll get back to you on this, as well as the paper, tomorrow.
