### 2022-06-28 wouterpeere/GHEtool#7 @wouterpeere
@nmstreethran do you have any updates on the paper or the examples?
