### 2022-06-28 wouterpeere/GHEtool#7 @nmstreethran
@wouterpeere My apologies for the delay in getting back to you. I'll get the review completed ASAP.

I have some additional comments on the tests:

1. Can you clarify the need to use `--pyargs GHEtool` for testing? When I installed the development version from TestPyPI and ran the tests with the `pyargs` flag, no tests run. All tests run when it is omitted.

2. I installed the development version just now and ran the tests again. I found a new issue: the `test_dynamicRb` test fails (I've included the test output below). Numpy was updated to v1.23.0 six days ago and they have expired the `asscalar` function's deprecation (see <https://github.com/numpy/numpy/releases/tag/v1.23.0>). You'll need to restrict the Numpy version to <= v1.22.4 in `setup.cfg`. (It may be easier to just use a newer Python version with the latest version of pygfunction and the development version of CoolProp though, but I'll leave this to you.)

<summary>Pytest output</summary>

```
=============================================== test session starts ================================================
rootdir: /home/nms/Downloads/ghetool
collected 9 items

GHEtool/GHEtool/test/test_GHEtool.py ........F                                                               [100%]

===================================================== FAILURES =====================================================
__________________________________________________ test_dynamicRb __________________________________________________

borefield = <GHEtool.GHEtool.Borefield object at 0x557355916f80>

    def test_dynamicRb(borefield):
        fluidData = FluidData(0.2, 0.568, 998, 4180, 1e-3)
        pipeData = PipeData(1, 0.015, 0.02, 0.4, 0.05, 0.075, 2)
        borefield.setFluidParameters(fluidData)
        borefield.setPipeParameters(pipeData)
>       assert round(borefield.size(100, useConstantRb=False), 2) == 52.57

GHEtool/GHEtool/test/test_GHEtool.py:103:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _
env/lib/python3.8/site-packages/GHEtool/GHEtool.py:528: in size
    result = self.sizeL2(H_init, quadrantSizing) if L2Sizing else self.sizeL3(H_init, quadrantSizing)
env/lib/python3.8/site-packages/GHEtool/GHEtool.py:583: in sizeL2
    quadrant1 = sizeQuadrant1()
env/lib/python3.8/site-packages/GHEtool/GHEtool.py:552: in sizeQuadrant1
    return self._Carcel  # size
env/lib/python3.8/site-packages/GHEtool/GHEtool.py:501: in _Carcel
    L = (self.qh * self._Rb + self.qh * Rh + self.qm * Rcm + self.qpm * Rpm) / abs(self.Tf - self.Tg)
env/lib/python3.8/site-packages/GHEtool/GHEtool.py:414: in _Rb
    return self.calculateRb()
env/lib/python3.8/site-packages/GHEtool/GHEtool.py:432: in calculateRb
    return gt.pipes.borehole_thermal_resistance(pipe, self.mfr, self.Cp)
env/lib/python3.8/site-packages/pygfunction/pipes.py:1870: in borehole_thermal_resistance
    a_out = np.asscalar(
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

attr = 'asscalar'

    def __getattr__(attr):
        # Warn for expired attributes, and return a dummy function
        # that always raises an exception.
        try:
            msg = __expired_functions__[attr]
        except KeyError:
            pass
        else:
            warnings.warn(msg, DeprecationWarning, stacklevel=2)

            def _expired(*args, **kwds):
                raise RuntimeError(msg)

            return _expired

        # Emit warnings for deprecated attributes
        try:
            val, msg = __deprecated_attrs__[attr]
        except KeyError:
            pass
        else:
            warnings.warn(msg, DeprecationWarning, stacklevel=2)
            return val

        # Importing Tester requires importing all of UnitTest which is not a
        # cheap import Since it is mainly used in test suits, we lazy import it
        # here to save on the order of 10 ms of import time for most users
        #
        # The previous way Tester was imported also had a side effect of adding
        # the full `numpy.testing` namespace
        if attr == 'testing':
            import numpy.testing as testing
            return testing
        elif attr == 'Tester':
            from .testing import Tester
            return Tester

>       raise AttributeError("module {!r} has no attribute "
                             "{!r}".format(__name__, attr))
E       AttributeError: module 'numpy' has no attribute 'asscalar'

env/lib/python3.8/site-packages/numpy/__init__.py:311: AttributeError
============================================= short test summary info ==============================================
FAILED GHEtool/GHEtool/test/test_GHEtool.py::test_dynamicRb - AttributeError: module 'numpy' has no attribute 'as...
=========================================== 1 failed, 8 passed in 1.71s ============================================
```
