### 2022-06-29 wouterpeere/GHEtool#7 @wouterpeere
@nmstreethran Thank you for the comments

> 1. Can you clarify the need to use `--pyargs GHEtool` for testing? When I installed the development version from TestPyPI and ran the tests with the `pyargs` flag, no tests run. All tests run when it is omitted.

Personally, for me this flag is needed, otherwise no tests would run ... I found some explanation on it [here](https://docs.pytest.org/en/7.1.x/example/pythoncollection.html#interpreting-cmdline-arguments-as-python-packages).
However, I noticed that in v2.0.3rc1 (from TestPyPi) the test weren't actually installed (they are not embedded in the package). I now uploaded v2.0.3rc2 to testPyPi, which has the tests embedded. Can you please delete everything from v2.0.3rc1 (including the test folder) and install the newer version. This solved the problem for me ...

> 2\. I installed the development version just now and ran the tests again. I found a new issue: the `test_dynamicRb` test fails (I've included the test output below). Numpy was updated to v1.23.0 six days ago and they have expired the `asscalar` function's deprecation (see https://github.com/numpy/numpy/releases/tag/v1.23.0). You'll need to restrict the Numpy version to <= v1.22.4 in `setup.cfg`. (It may be easier to just use a newer Python version with the latest version of pygfunction and the development version of CoolProp though, but I'll leave this to you.)

I noticed it myself. I think that, since many people will be willing to upgrade to a higher version of Numpy, your latest suggestion is better. I will change the installation instructions in the README.
