### 2022-06-30 wouterpeere/GHEtool#7 @nmstreethran
Some minor general comments: some words should be capitalised - Python (in paper and README), Boydens Engineering (in paper)
