### 2022-06-30 wouterpeere/GHEtool#7 @wouterpeere
> * installed GHEtool v2.0.3rc2 from Test PyPI
> * installed CoolProp v6.4.2.dev0 from Test PyPI
> * installed pygfunction v2.1.0

Perfect, I will use this in the README.

> Regarding the GUI, there is a precompiled EXE available. Is there a way users can compile it themselves? Is it possible to run it on platforms other than Windows?

For now, but that is the expertise of @tblanke, only Windows is possible, but he was working on making it work on other platforms. Now, users are not able to compile it themselves, but I guess, @tblanke, if you include a script for it (and add some documentation to the README) they will be able to? I will make a separate issue for this!
