### 2022-06-30 wouterpeere/GHEtool#7 @wouterpeere
> Regarding the paper, I found some issues with the bibliography file. I can open a PR to fix this if you prefer. I also noticed that there are two copies of the paper and bib file, in the root and the `JOSS` folder. Which ones are the main files? (are they symlinked?)

I duplicated them, since I was struggling to get it working. The one in the root is the correct one (I suppose).
What are the problems with the bibliography? If you want, you can use a PR to fix these issues, if that is not too much trouble for you.
