### 2022-06-30 0576c06
Include tests in build Corrected url in paper Python with capital letters Update README with new installation instructions
