+++ paper.md
@@ -26,7 +26,7 @@ bibliography: paper.bib
 
 # Summary
 
-GHEtool is a python package that contains all the functionalities needed to deal with borefield design. It is developed for both researchers and practitioners.
+GHEtool is a Python package that contains all the functionalities needed to deal with borefield design. It is developed for both researchers and practitioners.
 The core of this package is the automated sizing of borefield under different conditions.
 The sizing of a borefield is typically slow due to the high complexity of the mathematical background. Because this tool has a lot of precalculated data, GHEtool can size a borefield in the order of tenths of milliseconds.
 This sizing typically takes the order of minutes. Therefore, this tool is suited for being implemented in typical workflows where iterations are required.
@@ -44,7 +44,7 @@ Therefore, much interest exists from academia and practitioners in this field.
 One of the main challenges in this domain is to size a borefield, which is critical since borefields have a very high investment cost, so we want them to be as small as possible to increase their (economic) feasibility.
 In all research related to optimising borefield loads, studying sensitivity on borefield configurations, and integrating seasonal thermal energy storage in fifth-generation district heating etc. borefield sizing plays a central role.
 
-GHEtool is a python package with a graphical counterpart centred around borefield sizing and the evaluation of temperature evolution. 
+GHEtool is a Python package with a graphical counterpart centred around borefield sizing and the evaluation of temperature evolution. 
 
 The main advantage of GHEtool is that the ground response functions (gfunctions [@gfunctions]), which are needed to size a borefield, are already precalculated for more than 140 000 borefield – ground combinations.
 Due to this precalculation, one can save from a couple of seconds to several minutes to calculate the gfunction for even one borefield. On the other hand, the tool provides the option to provide your gfunction data for your own custom borefield configuration. 
@@ -64,7 +64,7 @@ Multiple tools are available for sizing borefields:
 - EED and GLHEPRO are commercial, standalone tools for borefields. They allow the user, a.o., to plot the mean fluid and borehole wall temperature evolution [@EED; @GLHEPRO]. Furthermore, the user can size a borefield [@EED; @GLHEPRO]. The automated sizing in EED sometimes leads to an undersized borefield [@Peere]. GLHEPRO does not have this problem since it calculates the temperature in every month [@GLHEPRO].
 - geoSIM is a free tool for simulating and sizing borefields but only with one particular type of tubing (geoKOAX). It is also standalone [@geoSIM].
 - Ground Loop Design software (GLDTM) ‘is the world’s leading commercial GHX software design tool’ [@igshpa]. It allows the user, a.o., to plot the mean fluid and borehole wall temperature evolution. Furthermore, the user can size the borefield.
-- A commonly used package in the borefield domain is pygfunction. Pygfunction is an open-source python package to calculate the thermal response factors of the ground. This function forms the basis of many borefield simulation and sizing programs [@pygfunction]. With this python package, the data of GHEtool was precalculated.
+- A commonly used package in the borefield domain is pygfunction. Pygfunction is an open-source Python package to calculate the thermal response factors of the ground. This function forms the basis of many borefield simulation and sizing programs [@pygfunction]. With this Python package, the data of GHEtool was precalculated.
 
 From the mentioned tools, only EED, GLHEPRO, geoSIM and GLDTM can size borefields, but none of them is open source nor easy to be integrated into your workflow since they are standalone. From all the borefield sizing tools listed above, only geoSIM is free, but it only works with one specific type of borehole tubing [@geoSIM].
 (Bernier, 2019) has reviewed exhaustive literature of all existing borefield sizing methods and tools [@Bernier].
