+++ paper.bib
@@ -49,7 +49,8 @@
 @article{IEA,
   title={GLOBAL STATUS REPORT: Towards a zero-emission, efficient, and resilient buildings and construction sector},
   author={Abergel, Thibaut and Dean, Brian and Dulac, John and Hamilton, Ian and Wheeler, T},
-  journal={Global Alliance for Buildings and Construction. Available online: https://www. worldgbc. org/sites/default/files/2018\% 20GlobalABC\% 20Global\% 20Status\% 20Report (accessed on 28 February 2022)},
+  journal={Global Alliance for Buildings and Construction.},
+  note= {Available online: https://www.worldgbc.org/sites/default/files/2018\%20GlobalABC\%20Global\%20Status\%20Report (accessed on 28 February 2022)},
   year={2018}},
 
 @misc{geoSIM,
