+++ MANIFEST.IN
@@ -1,1 +1,2 @@
 include GHEtool/Data/*.pickle
+include GHEtool/test/*.py
