+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = GHEtool
-version = 2.0.3rc1
+version = 2.0.3rc3
 author = Wouter Peere
 author_email = wouter.peere@kuleuven.be
 description = Python package for borefield sizing
@@ -20,7 +20,7 @@ install_requires =
     numpy>=1.20.2
 	scipy>=1.6.2
 	matplotlib>=3.4.1
-	pygfunction==1.1.2
+	pygfunction>=2.1.0
 	openpyxl>=3.0.7
 	pandas>=1.2.4
 	pytest>=7.1.2
