### 2022-06-30 wouterpeere/GHEtool#7 @wouterpeere
> I think the instructions for compiling the GUI would be useful. But if it's too much work, then you can just make it clear in the README that the compiled GUI is Windows-only.

I will leave this up to @tblanke.

> Regarding 'boydens engineering', if that's the legal name then you can just keep it that way. It's capitalised in the README though, so maybe you can change that?

I will change it!

> In the bibliography, the IEA report's link is broken, and it uses the `note` field for the conference location instead of the `location` field, among others. I also found the conference proceedings online and I think it would be useful to add those links in the bibliography. I'll create a PR and you can review the changes.

Thanks!
