+++ paper.bib
@@ -1,66 +1,86 @@
-@INPROCEEDINGS{GLHEPRO,
-    author = {Jeffrey D. Spitler},
-    title = {{GLHEPRO}-- A design tool for commercial building ground loop heat exchangers},
-    booktitle = {Proceedings of the fourth international heat pumps in cold climates conference},
-    year = {2000},
-    note={Aylmer, Québec. August 17-18, 2000.}},
+@inproceedings{GLHEPRO,
+  author = {Jeffrey D. Spitler},
+  title = {{GLHEPRO} -- {A Design Tool For Commercial Building Ground Loop Heat Exchangers}},
+  booktitle = {Proceedings of the Fourth International Heat Pumps in Cold Climates Conference},
+  date = {2000-08},
+  location = {Aylmer, QC, Canada},
+  url = {https://hvac.okstate.edu/sites/default/files/pubs/papers/2000/08-HPCC_GLHEPRO.pdf},
+  urldate = {2022-06-30}
+}
 
 @article{Bernier,
-	title = "A review of vertical ground heat exchanger sizing tools including an inter-model comparison",
-	journal = "Renewable and Sustainable Energy Reviews",
-	volume = "110",
-	pages = "247 - 265",
-	year = "2019",
-	issn = "1364-0321",
-	doi = "10.1016/j.rser.2019.04.045",
-	url = "http://www.sciencedirect.com/science/article/pii/S1364032119302576",
-	author = "Mohammadamin Ahmadfard and Michel Bernier",
-	keywords = "Geothermal energy, Sizing tools, Vertical ground heat exchangers, Inter-model comparison, Test cases"},
+  title = "A review of vertical ground heat exchanger sizing tools including an inter-model comparison",
+  journal = "Renewable and Sustainable Energy Reviews",
+  volume = "110",
+  pages = "247 - 265",
+  year = "2019",
+  issn = "1364-0321",
+  doi = "10.1016/j.rser.2019.04.045",
+  url = "http://www.sciencedirect.com/science/article/pii/S1364032119302576",
+  author = "Mohammadamin Ahmadfard and Michel Bernier",
+  keywords = "Geothermal energy, Sizing tools, Vertical ground heat exchangers, Inter-model comparison, Test cases"
+}
 
-@inproceedings{gfunctions,
-  title={Thermal analysis of heat extraction boreholes},
-  author={Per Eskilson},
-  year={1987},
-  booktitle    = {PhD thesis, Dep. of Mathematical Physics, University of Lund, Sweden}},
+@thesis{gfunctions,
+  title = {{Thermal Analysis} of {Heat Extraction Boreholes}},
+  author = {Per Eskilson},
+  date = {1987-06},
+  institution = {Dep. of Mathematical Physics, University of Lund},
+  type = {PhD Thesis},
+  location = {Lund, Sweden},
+  url = {https://www.buildingphysics.com/download/Eskilson1987.pdf},
+  urldate = {2022-06-30}
+}
 
 @inproceedings{pygfunction,
-  title={pygfunction: an open-source toolbox for the evaluation of thermal response factors for geothermal borehole fields},
-  author={Massimo Cimmino},
-  year={2018},
-  booktitle    = {Proceedings of eSim 2018, the 10th conference of IBPSA-Canada},
-  note = {Montréal, QC, Canada, May 9-10}},
+  title = {{pygfunction}: {an} open-source toolbox for the evaluation of thermal response factors for geothermal borehole fields},
+  author = {Massimo Cimmino},
+  date = {2018-05},
+  booktitle = {Proceedings of eSim 2018, the 10th conference of IBPSA-Canada},
+  location = {Montréal, QC, Canada},
[…]
