+++ paper.md
@@ -38,7 +38,7 @@ A setup to install the GUI at the user-defined place is also implemented and ava
 # Statement of need
 
 The building sector uses 36 % of global energy and is responsible for 39 % of energy-related CO2 emissions when upstream power generation is included, with more than half of this energy used by space heating, space cooling and water heating [@IEA].
-This sector should decarbonize to reach the climate goals in 2050. One of the promising ways to achieve this goal is by using shallow geothermal energy, thereby storing heat/cold in borefields, enabling seasonal thermal energy storage.
+In order to reach the climate goals in 2050, as set by the European Green Deal, the building sector should decarbonise [@greenDeal]. One of the promising ways to achieve this goal is by using shallow geothermal energy, thereby storing heat/cold in borefields, enabling seasonal thermal energy storage.
 Therefore, much interest exists from academia and practitioners in this field.
 
 One of the main challenges in this domain is to size a borefield, which is critical since borefields have a very high investment cost, so we want them to be as small as possible to increase their (economic) feasibility.
@@ -63,7 +63,7 @@ Multiple tools are available for sizing borefields:
 
 - EED and GLHEPRO are commercial, standalone tools for borefields. They allow the user, a.o., to plot the mean fluid and borehole wall temperature evolution [@EED; @GLHEPRO]. Furthermore, the user can size a borefield [@EED; @GLHEPRO]. The automated sizing in EED sometimes leads to an undersized borefield [@Peere]. GLHEPRO does not have this problem since it calculates the temperature in every month [@GLHEPRO].
 - geoSIM is a free tool for simulating and sizing borefields but only with one particular type of tubing (geoKOAX). It is also standalone [@geoSIM].
-- Ground Loop Design software (GLDTM) ‘is the world’s leading commercial GHX software design tool’ [@igshpa]. It allows the user, a.o., to plot the mean fluid and borehole wall temperature evolution. Furthermore, the user can size the borefield.
+- Ground Loop Design software (GLDTM) ‘is the world’s leading commercial GHX software design tool’ [@igshpa; @GLDTM]. It allows the user, a.o., to plot the mean fluid and borehole wall temperature evolution. Furthermore, the user can size the borefield.
 - A commonly used package in the borefield domain is pygfunction. Pygfunction is an open-source Python package to calculate the thermal response factors of the ground. This function forms the basis of many borefield simulation and sizing programs [@pygfunction]. With this Python package, the data of GHEtool was precalculated.
 
 From the mentioned tools, only EED, GLHEPRO, geoSIM and GLDTM can size borefields, but none of them is open source nor easy to be integrated into your workflow since they are standalone. From all the borefield sizing tools listed above, only geoSIM is free, but it only works with one specific type of borehole tubing [@geoSIM].
