+++ paper.bib
@@ -59,8 +59,21 @@
 	note = {Available online: https://geokoax.com/en/our-products/simulation/ (accessed on 10 February 2022)},
 	url = {https://geokoax.com/en/our-products/simulation/}},
 
+@misc{GLDTM,
+	author = {{Thermal Dynamics Inc.}},
+	title = {{GLD SOFTWARE}},
+	note = {Available online: https://groundloopdesign.com/ (accessed on 19 July 2022)},
+	url = {https://groundloopdesign.com/}},
+
 @misc{igshpa,
-	author = {{Igshpa}},
+	author = {{igshpa}},
 	title = {{Software – Tools to design and support your Ground Source Heat Pump Installation}},
 	note = {Available online: https://igshpa.org/software/ (accessed on 10 February 2022)},
-	url = {https://igshpa.org/software/}}
+	url = {https://igshpa.org/software/}},
+
+@misc{greenDeal,
+	author = {{The European Commission}},
+	title = {{A European Green Deal}},
+	subtitle = {{Striving to be the first climate-neutral continent}},
+	note = {Available online: https://ec.europa.eu/info/strategy/priorities-2019-2024/european-green-deal_en (accessed on 19 July 2022)},
+	url = {https://ec.europa.eu/info/strategy/priorities-2019-2024/european-green-deal_en}},
