### 2022-07-19 2020fb8
Revert "Update paper"
This reverts commit e92235094ab46df0f54819618a80b21570ac3462.
