+++ paper.bib
@@ -83,4 +83,19 @@
   title = {{Software} - {Tools} to design and support your {Ground Source Heat Pump Installation}},
   url = {https://igshpa.org/software/},
   urldate = {2022-06-30}
+},
+
+@online{GLDTM,
+  author = {{Thermal Dynamics Inc.}},
+  title = {{GLD SOFTWARE}},
+  url = {https://groundloopdesign.com/},
+  urldate = {2022-07-22}
+},
+
+@misc{greenDeal,
+	author = {{The European Commission}},
+	title = {{A European Green Deal}},
+	subtitle = {{Striving to be the first climate-neutral continent}},
+	note = {Available online: https://ec.europa.eu/info/strategy/priorities-2019-2024/european-green-deal_en (accessed on 19 July 2022)},
+	url = {https://ec.europa.eu/info/strategy/priorities-2019-2024/european-green-deal_en}
 }
