### 2022-07-19 8faccce
Change .gitignore Remove duplicates of paper
