### 2022-07-25 41e165c
Implemented variable ground temperature
