+++ GHEtool/GHEtool.py
@@ -48,7 +48,7 @@ class Borefield:
         hourlyLoadArray.append(temp)
 
     __slots__ = 'baseloadHeating', 'baseloadCooling', 'H', 'H_init', 'B', 'N_1', 'N_2', 'Rb', 'k_s', 'Tg', 'ty', 'tm', \
-                'td', 'time', 'hourlyHeatingLoad', 'H_max', \
+                'td', 'time', 'hourlyHeatingLoad', 'H_max', 'useConstantTg', 'flux', \
                 'hourlyCoolingLoad', 'numberOfBoreholes', 'borefield', 'customGfunction', 'costInvestment', \
                 'lengthPeak', 'th', 'Tf_H', 'Tf_C', 'limitingQuadrant', 'monthlyLoad', 'monthlyLoadHeating', \
                 'monthlyLoadCooling', 'peakHeating', 'imbalance', 'qa', 'Tf', 'qm', 'qh', 'qpm', 'tcm', 'tpm', \
@@ -59,7 +59,8 @@ class Borefield:
                 'monthlyLoadHeatingExternal', 'monthlyLoadCoolingExternal', 'hourlyHeatingLoadExternal', \
                 'hourlyCoolingLoadExternal', 'hourlyHeatingLoadOnTheBorefield', 'hourlyCoolingLoadOnTheBorefield', \
                 'k_f', 'mfr', 'Cp', 'mu', 'rho', 'useConstantRb', 'h_f', 'R_f', 'R_p', 'printing', 'combo', \
-                'r_in', 'r_out', 'k_p', 'D_s', 'r_b', 'numberOfPipes', 'epsilon', 'k_g', 'pos', 'D'
+                'r_in', 'r_out', 'k_p', 'D_s', 'r_b', 'numberOfPipes', 'epsilon', 'k_g', 'pos', 'D', \
+                'L2sizing', 'L3sizing', 'L4sizing', 'quadrantSizing', 'H_init'
 
     def __init__(self, simulationPeriod: int = 20, numberOfBoreholes: int = None, peakHeating: list = None,
                  peakCooling: list = None,
@@ -84,6 +85,8 @@ class Borefield:
         self.thresholdWarningShallowField: int = 50
         # parameter that determines whether or not the Rb-value should be altered in the optimisation
         self.useConstantRb = True
+        # parameter that determines whether or not the Tg should be calculated with a heat flux or without
+        self.useConstantTg = True
 
         self.H_max: float = 350  # max threshold for interpolation (will with first sizing)
 
@@ -155,6 +158,7 @@ class Borefield:
         self.Rb = 0.  # effective borehole thermal resistance mK/W
         self.N_1 = 0  # number of boreholes in one direction #
         self.N_2 = 0  # number of boreholes in the other direction #
+        self.flux = 0.  # geothermal heat flux W/m2
         self.numberOfBoreholes = 0  # number of total boreholes #
 
         # initiate fluid parameters
@@ -413,6 +417,24 @@ class Borefield:
         # calculate Rb*
         return self.calculateRb()
 
+    def _Tg(self, H: float = None) -> float:
+        """
+        This function gives back the ground temperature
+        When useConstantTg is False, the thermal heat flux is taken into account.
+
+        :param H: depth of the field (optional)
+        :return: ground temperature
+        """
+        if self.useConstantTg:
+            return self.Tg
+
+        # check if specific depth is given
+        if H is None:
+            H is self.H
+
+        # geothermal gradient is equal to the geothermal heat flux divided by the thermal conductivity
+        return self.Tg + H * self.flux / self.k_s
+
     def calculateRb(self) -> float:
         """
[…]
