### 2022-07-25 3aab708
Implemented sizingsetup and typos
