### 2022-07-26 5353a5c
Paper changes, added example for the speed comparison with and without precalculated data Small restructuring, typo's
