+++ paper.bib
@@ -98,4 +98,49 @@
 	subtitle = {{Striving to be the first climate-neutral continent}},
 	note = {Available online: https://ec.europa.eu/info/strategy/priorities-2019-2024/european-green-deal_en (accessed on 19 July 2022)},
 	url = {https://ec.europa.eu/info/strategy/priorities-2019-2024/european-green-deal_en}
+},
+
+@article{investmentcost,
+title = {Hybrid ground-source heat pump system with active air source regeneration},
+journal = {Energy Conversion and Management},
+volume = {90},
+pages = {230-237},
+year = {2015},
+issn = {0196-8904},
+doi = {https://doi.org/10.1016/j.enconman.2014.11.009},
+url = {https://www.sciencedirect.com/science/article/pii/S0196890414009650},
+author = {K. Allaerts and M. Coomans and R. Salenbien}},
+
+@article{sensitivity,
+title = {A review on energy piles design, sizing and modelling},
+journal = {Energy},
+volume = {122},
+pages = {390-407},
+year = {2017},
+issn = {0360-5442},
+doi = {https://doi.org/10.1016/j.energy.2017.01.097},
+url = {https://www.sciencedirect.com/science/article/pii/S0360544217301044},
+author = {Jevgeni Fadejev and Raimo Simson and Jarek Kurnitski and Fariborz Haghighat}},
+
+@Article{loads,
+AUTHOR = {Cupeiro Figueroa, Iago and Cimmino, Massimo and Helsen, Lieve},
+TITLE = {A Methodology for Long-Term Model Predictive Control of Hybrid Geothermal Systems: The Shadow-Cost Formulation},
+JOURNAL = {Energies},
+VOLUME = {13},
+YEAR = {2020},
+NUMBER = {23},
+ARTICLE-NUMBER = {6203},
+URL = {https://www.mdpi.com/1996-1073/13/23/6203},
+ISSN = {1996-1073},
+DOI = {10.3390/en13236203}},
+
+@thesis{districtHeating,
+  title = {Design and Assessment of low-carbon Residential District Concepts with (Collective) Seasonal Thermal Energy Storage},
+  author = {Louis Hermans, Robin Haesen},
+  date = {2021},
+  institution = {Dep. of Mechanical Engineering, KU Leuven (university of Leuven)},
+  type = {Master Thesis},
+  location = {Leuven, Belgium},
+  url = {https://www.scriptiebank.be/sites/default/files/thesis/2021-10/Masters_thesis_RobinHaesen_LouisHermans.pdf},
+  urldate = {2022-07-26}
 }
