### 2022-07-26 wouterpeere/GHEtool#7 @wouterpeere
I implemented the other changes and added some more references to the paper.
