+++ paper.bib
@@ -136,7 +136,7 @@ DOI = {10.3390/en13236203}},
 
 @thesis{districtHeating,
   title = {Design and Assessment of low-carbon Residential District Concepts with (Collective) Seasonal Thermal Energy Storage},
-  author = {Louis Hermans, Robin Haesen},
+  author = {Hermans, Louis and Haesen, Robin},
   date = {2021},
   institution = {Dep. of Mechanical Engineering, KU Leuven (university of Leuven)},
   type = {Master Thesis},
