### 2022-07-27 wouterpeere/GHEtool#7 @nmstreethran
@wouterpeere Thanks! Can you add a link to the speed comparison script in the README?

> I'm doubting whether referencing this article is an added value for this paper ...

I think the script you added is enough to verify the speed claim, so you don't need to reference that article.

In the last sentence of the Statement of need:

> ... to play an essential role in the energy transition and of the built environment.

I'm not sure if the "and" should be there...

The last sentence of the Comparison with existing tools should be changed to the following so that you don't repeat the author's name:

`@Bernier has reviewed exhaustive literature of all existing borefield sizing methods and tools.`

The CO2 subscript didn't seem to work. Can you try `CO~2~` instead (as stated in the [Pandoc manual](https://pandoc.org/MANUAL.html#superscripts-and-subscripts)?
