### 2022-07-28 33f84b0
Changes to paper + added speed comparison to the README
