### 2022-07-28 wouterpeere/GHEtool#7 @wouterpeere
@nmstreethran the changes have been implemented!
I added another table under the functionalities subsection for the comparison documents (both L2/L3 sizing and the speed comparison).
