### 2022-07-28 wouterpeere/GHEtool#7 @nmstreethran
Thanks @wouterpeere! It's nice to have a separate table for the comparison documents.

That's it from me. I'll close this issue now and update the editor.
