### 2022-07-28 1ce2eab
Changes to self._Tg() and typo's
