+++ GHEtool/GHEtool.py
@@ -223,7 +223,7 @@ class Borefield:
         # calculate number of boreholes
         self.setNumberOfBoreholes(numberOfBoreholes if numberOfBoreholes is not None else 1)
 
-        # set a custom bore field
+        # set a custom borefield
         self.borefield = None
         self.setBorefield(borefield)
 
@@ -249,7 +249,7 @@ class Borefield:
 
     def setBorefield(self, borefield=None) -> None:
         """
-        This function sets the bore field configuration. When no input, an empty array of length N_1 * N_2 will be made.
+        This function sets the borefield configuration. When no input, an empty array of length N_1 * N_2 will be made.
         :return None
         """
         if borefield is None:
@@ -433,7 +433,7 @@ class Borefield:
 
         # check if specific depth is given
         if H is None:
-            H is self.H
+            H = self.H
 
         # geothermal gradient is equal to the geothermal heat flux divided by the thermal conductivity
         return self.Tg + H * self.flux / self.k_s
@@ -449,7 +449,7 @@ class Borefield:
             print("Please make sure you set al the pipe and fluid data.")
             raise ValueError
 
-        # initiate temporary bore field
+        # initiate temporary borefield
         borehole = gt.boreholes.Borehole(self.H, self.D, self.r_b, 0, 0)
         # initiate pipe
         pipe = gt.pipes.MultipleUTube(self.pos, self.r_in, self.r_out, borehole, self.k_s, self.k_g,
@@ -531,7 +531,7 @@ class Borefield:
         return self.H
 
     def sizingSetup(self, H_init: float = 100, useConstantRb: bool = None, useConstantTg: bool = None, quadrantSizing: int = 0,
-                    L2sizing: bool = None, L3sizing: bool = None, L4sizing: bool = None) -> None:
+                    L2sizing: bool = None, L3sizing: bool = None) -> None:
         """
         This function sets the options for the sizing function.
         * The L2 sizing is the one explained in (Peere et al., 2021) and is the quickest method (it uses 3 pulses)
@@ -544,12 +544,11 @@ class Borefield:
         :param quadrantSizing: differs from 0 when a sizing in a certain quadrant is desired
         :param L2sizing: true if a sizing with the L2 method is needed
         :param L3sizing: true if a sizing with the L3 method is needed
-        :param L4sizing: true if a sizing with the L4 method is needed
         :return: None
         """
 
         # check if just one sizing is given
-        if np.sum([L2sizing if L2sizing is not None else 0, L3sizing if L3sizing is not None else 0, L4sizing if L4sizing is not None else 0]) > 1:
+        if np.sum([L2sizing if L2sizing is not None else 0, L3sizing if L3sizing is not None else 0]) > 1:
             raise ValueError("Please check if just one sizing method is chosen!")
 
         # set variables
@@ -565,25 +564,25 @@ class Borefield:
             self.L2sizing = L2sizing
         if L3sizing is not None:
             self.L3sizing = L3sizing
-        if L4sizing is not None:
-            self.L4sizing = L4sizing
+        # if L4sizing is not None:
+        #     self.L4sizing = L4sizing
 
     def size(self, H_init: float = 100, useConstantRb: bool = None, useConstantTg: bool = None,
-             L2sizing: bool = None, L3sizing: bool = None, L4sizing: bool = None, quadrantSizing: int = None) -> float:
+             L2sizing: bool = None, L3sizing: bool = None, quadrantSizing: int = None) -> float:
         """
         This function sizes the borefield. It lets the user chose between three sizing methods.
         * The L2 sizing is the one explained in (Peere et al., 2021) and is the quickest method (it uses 3 pulses)
[…]
