### 2022-08-01 ce32f8e
Bugfix in var temperature
