### 2022-08-02 ac856c7
New setup.cfg for new implementations
