### 2022-08-02 377df9a
Implemented framework for L4-sizing Implemented function for setting hourly loads Changed name of speed comparison validation Added new document for speed comparison for the different implementations of the L4 method (TBC)
