+++ GHEtool/GHEtool.py
@@ -532,7 +532,7 @@ class Borefield:
         return self.H
 
     def sizingSetup(self, H_init: float = 100, useConstantRb: bool = None, useConstantTg: bool = None, quadrantSizing: int = 0,
-                    L2sizing: bool = None, L3sizing: bool = None) -> None:
+                    L2sizing: bool = None, L3sizing: bool = None, L4sizing: bool = None) -> None:
         """
         This function sets the options for the sizing function.
         * The L2 sizing is the one explained in (Peere et al., 2021) and is the quickest method (it uses 3 pulses)
@@ -545,11 +545,12 @@ class Borefield:
         :param quadrantSizing: differs from 0 when a sizing in a certain quadrant is desired
         :param L2sizing: true if a sizing with the L2 method is needed
         :param L3sizing: true if a sizing with the L3 method is needed
+        :param L4sizing: true if a sizing with the L4 method is needed
         :return: None
         """
 
         # check if just one sizing is given
-        if np.sum([L2sizing if L2sizing is not None else 0, L3sizing if L3sizing is not None else 0]) > 1:
+        if np.sum([L2sizing if L2sizing is not None else 0, L3sizing if L3sizing is not None else 0, L4sizing if L4sizing is not None else 0]) > 1:
             raise ValueError("Please check if just one sizing method is chosen!")
 
         # set variables
@@ -565,15 +566,16 @@ class Borefield:
             self.L2sizing = L2sizing
         if L3sizing is not None:
             self.L3sizing = L3sizing
-        # if L4sizing is not None:
-        #     self.L4sizing = L4sizing
+        if L4sizing is not None:
+            self.L4sizing = L4sizing
 
     def size(self, H_init: float = 100, useConstantRb: bool = None, useConstantTg: bool = None,
-             L2sizing: bool = None, L3sizing: bool = None, quadrantSizing: int = None) -> float:
+             L2sizing: bool = None, L3sizing: bool = None, L4sizing: bool = None, quadrantSizing: int = None) -> float:
         """
         This function sizes the borefield. It lets the user chose between three sizing methods.
         * The L2 sizing is the one explained in (Peere et al., 2021) and is the quickest method (it uses 3 pulses)
         * The L3 sizing is a more general approach which is slower but more accurate (it uses 24 pulses/year)
+        * The L4 sizing is the most exact one, since it uses hourly data (8760 pulses/year)
 
         Please note that the changes sizing setup changes here are not saved! Use self.setupSizing for this.
         (e.g. if you size by putting the constantTg param to True but it was False, if you plot the results afterwards
@@ -584,6 +586,7 @@ class Borefield:
         :param useConstantTg: true if a constant Tg value should be used (the geothermal flux is neglected)
         :param L2sizing: true if a sizing with the L2 method is needed
         :param L3sizing: true if a sizing with the L3 method is needed
+        :param L4sizing: true if a sizing with the L4 method is needed
         :param quadrantSizing: differs from 0 when a sizing in a certain quadrant is desired
         :return: borefield depth
         """
@@ -592,19 +595,19 @@ class Borefield:
 
         # run the sizing setup
         self.sizingSetup(H_init=H_init, useConstantRb=useConstantRb, useConstantTg=useConstantTg,
-                         L2sizing=L2sizing, L3sizing=L3sizing, quadrantSizing=quadrantSizing)
+                         L2sizing=L2sizing, L3sizing=L3sizing, L4sizing=L4sizing, quadrantSizing=quadrantSizing)
 
[…]
