### 2022-08-02 193d07a
Added development and application papers of GHEtool to the README
