+++ paper.bib
@@ -100,6 +100,14 @@
 	url = {https://ec.europa.eu/info/strategy/priorities-2019-2024/european-green-deal_en}
 },
 
+@misc{greenDeal2,
+	author = {{The European Commission}},
+	title = {{European Green Deal: Commission proposes to boost renovation and decarbonisation of buildings}},
+	year = {2021},
+	note = {Available online: https://ec.europa.eu/commission/presscorner/detail/en/IP_21_6683 (accessed on 3 August 2022)},
+	url = {https://ec.europa.eu/commission/presscorner/detail/en/IP_21_6683}
+},
+
 @article{investmentcost,
 title = {Hybrid ground-source heat pump system with active air source regeneration},
 journal = {Energy Conversion and Management},
