### 2022-08-03 a59a64e
Update paper.md
Typo, added extra space
