### 2022-08-04 908f292
Removed print statement in self._printTemperatureProfile()
