### 2022-08-04 4a6dfa6
Change formatting of the code to PEP8 standard
