+++ GHEtool/Examples/Import_Data.py
@@ -14,12 +14,12 @@ if __name__ == "__main__":
     borefield = Borefield()
 
     # set ground data in borefield
-    borefield.setGroundParameters(data)
+    borefield.set_ground_parameters(data)
 
     # load the hourly profile
-    borefield.loadHourlyProfile("Hourly_Profile.csv", header=True, separator=";", firstColumnHeating=True)
-    borefield.convertHourlyToMonthly()
+    borefield.load_hourly_profile("hourly_profile.csv", header=True, separator=";", first_column_heating=True)
+    borefield.convert_hourly_to_monthly()
 
     # size the borefield and plot the resulting temperature evolution
     borefield.size(100)
-    borefield.printTemperatureProfile()
+    borefield.print_temperature_profile()
