+++ GHEtool/Validation/Speed_Comparison.py
@@ -1,5 +1,5 @@
 """
-This document compares the speed of the L2 sizing method of (Peere et al., 2021) with and without the precalculated gfunction data.
+This document compares the speed of the l_2 sizing method of (Peere et al., 2021) with and without the precalculated gfunction data.
 This is done for two fields with different sizes. It shows that, specifically for the larger fields, the precalculated data is way faster.
 """
 
@@ -8,7 +8,7 @@ import time
 from GHEtool import Borefield, GroundData
 
 
-def test64boreholes():
+def test_64_boreholes():
     data = GroundData(110, 6, 3, 10, 0.2, 8, 8)
 
     # montly loading values
@@ -28,17 +28,17 @@ def test64boreholes():
     monthlyLoadCooling = list(map(lambda x: x * annualCoolingLoad, montlyLoadCoolingPercentage))  # kWh
 
     # create the borefield object
-    borefield = Borefield(simulationPeriod=20,
-                          peakHeating=peakHeating,
-                          peakCooling=peakCooling,
-                          baseloadHeating=monthlyLoadHeating,
-                          baseloadCooling=monthlyLoadCooling)
+    borefield = Borefield(simulation_period=20,
+                          peak_heating=peakHeating,
+                          peak_cooling=peakCooling,
+                          baseload_heating=monthlyLoadHeating,
+                          baseload_cooling=monthlyLoadCooling)
 
-    borefield.setGroundParameters(data)
+    borefield.set_ground_parameters(data)
 
     # set temperature boundaries
-    borefield.setMaxGroundTemperature(16)  # maximum temperature
-    borefield.setMinGroundTemperature(0)  # minimum temperature
+    borefield.set_max_ground_temperature(16)  # maximum temperature
+    borefield.set_min_ground_temperature(0)  # minimum temperature
 
     # size borefield
     t1 = time.time()
@@ -47,7 +47,7 @@ def test64boreholes():
 
     ### size without the precalculation
 
-    borefield.usePrecalculatedData = False
+    borefield.use_precalculated_data = False
     # size borefield
     t2 = time.time()
     depthCalculated = borefield.size(100)
@@ -59,58 +59,58 @@ def test64boreholes():
           round((depthCalculated - depthPrecalculated) / depthCalculated * 100, 3), "%.")
 
 
-def test10boreholes():
+def test_10_boreholes():
     data = GroundData(110, 6, 3, 10, 0.2, 2, 5)
 
-    # montly loading values
-    peakCooling = [0., 0, 3., 9., 13., 20., 43., 30., 16., 7., 0., 0.]  # Peak cooling in kW
-    peakHeating = [16., 14, 10., 5., 0., 0., 0., 0., 4, 8., 19., 13.]  # Peak heating in kW
+    # monthly loading values
+    peak_cooling = [0., 0, 3., 9., 13., 20., 43., 30., 16., 7., 0., 0.]  # Peak cooling in kW
+    peak_heating = [16., 14, 10., 5., 0., 0., 0., 0., 4, 8., 19., 13.]  # Peak heating in kW
 
     # annual heating and cooling load
-    annualHeatingLoad = 16 * 10 ** 3  # kWh
-    annualCoolingLoad = 24 * 10 ** 3  # kWh
+    annual_heating_load = 16 * 10 ** 3  # kWh
+    annual_cooling_load = 24 * 10 ** 3  # kWh
 
     # percentage of annual load per month (15.5% for January ...)
-    montlyLoadHeatingPercentage = [0.155, 0.148, 0.125, .099, .064, 0., 0., 0., 0.061, 0.087, 0.117, 0.144]
-    montlyLoadCoolingPercentage = [0.025, 0.05, 0.05, .05, .075, .1, .2, .2, .1, .075, .05, .025]
[…]
