+++ GHEtool/Validation/Sizing_With_Rb_Calculation.py
@@ -1,6 +1,6 @@
 """
 This document compares the sizing with a constant Rb*-value with sizing where the Rb*-value is being recalculated.
-For the test, the L2 sizing method is used.
+For the test, the l_2 sizing method is used.
 The comparison is based on speed and relative accuracy in the result.
 It is shown that the speed difference is significant, but so is the difference in the result. With a constant Rb* value, it is important that the initial depth is rather accurate.
 """
@@ -11,149 +11,149 @@ if __name__ == "__main__":
 
     from GHEtool import Borefield, GroundData, FluidData, PipeData
 
-    numberOfIterations = 50
-    maxValueCooling = 700
-    maxValueHeating = 800
+    number_of_iterations = 50
+    max_value_cooling = 700
+    max_value_heating = 800
 
     # initiate the arrays
-    resultsRbStatic = np.empty(numberOfIterations)
-    resultsRbDynamic = np.empty(numberOfIterations)
-    differenceResults = np.empty(numberOfIterations)
+    results_Rb_static = np.empty(number_of_iterations)
+    results_Rb_dynamic = np.empty(number_of_iterations)
+    difference_results = np.empty(number_of_iterations)
 
-    monthlyLoadCoolingArray = np.empty((numberOfIterations, 12))
-    monthlyLoadHeatingArray = np.empty((numberOfIterations, 12))
-    peakLoadCoolingArray = np.empty((numberOfIterations, 12))
-    peakLoadHeatingArray = np.empty((numberOfIterations, 12))
+    monthly_load_cooling_array = np.empty((number_of_iterations, 12))
+    monthly_load_heating_array = np.empty((number_of_iterations, 12))
+    peak_load_cooling_array = np.empty((number_of_iterations, 12))
+    peak_load_heating_array = np.empty((number_of_iterations, 12))
 
     # populate arrays with random values
-    for i in range(numberOfIterations):
+    for i in range(number_of_iterations):
         for j in range(12):
-            monthlyLoadCoolingArray[i, j] = np.random.randint(0, maxValueCooling)
-            monthlyLoadHeatingArray[i, j] = np.random.randint(0, maxValueHeating)
-            peakLoadCoolingArray[i, j] = np.random.randint(monthlyLoadCoolingArray[i, j], maxValueCooling)
-            peakLoadHeatingArray[i, j] = np.random.randint(monthlyLoadHeatingArray[i, j], maxValueHeating)
+            monthly_load_cooling_array[i, j] = np.random.randint(0, max_value_cooling)
+            monthly_load_heating_array[i, j] = np.random.randint(0, max_value_heating)
+            peak_load_cooling_array[i, j] = np.random.randint(monthly_load_cooling_array[i, j], max_value_cooling)
+            peak_load_heating_array[i, j] = np.random.randint(monthly_load_heating_array[i, j], max_value_heating)
 
     # initiate borefield model
     data = GroundData(100, 6, 3, 10, 0.2, 10, 12)  # ground data with an inaccurate guess of 100m for the depth of the borefield
-    fluidData = FluidData(0.2, 0.568, 998, 4180, 1e-3)
-    pipeData = PipeData(1, 0.015, 0.02, 0.4, 0.05, 0.075, 2)
+    fluid_data = FluidData(0.2, 0.568, 998, 4180, 1e-3)
+    pipe_data = PipeData(1, 0.015, 0.02, 0.4, 0.05, 0.075, 2)
 
-    # Montly loading values
-    peakCooling = [0., 0, 34., 69., 133., 187., 213., 240., 160., 37., 0., 0.]              # Peak cooling in kW
-    peakHeating = [160., 142, 102., 55., 0., 0., 0., 0., 40.4, 85., 119., 136.]             # Peak heating in kW
+    # Monthly loading values
[…]

[…]
