### 2022-08-05 8242d40
Changed the examples and README to include the ground volumetric heat capacity
