+++ GHEtool/Examples/effect_of_borehole_configuration.py
@@ -7,7 +7,7 @@ from GHEtool import *
 
 if __name__ == "__main__":
     # GroundData for an initial field of 11 x 11
-    data = GroundData(110, 6, 3, 10, 0.2, 11, 11)
+    data = GroundData(110, 6, 3, 10, 0.2, 11, 11, 2.4 * 10**6)
 
     # Monthly loading values
     peak_cooling = [0., 0, 34., 69., 133., 187., 213., 240., 160., 37., 0., 0.]              # Peak cooling in kW
