+++ GHEtool/Validation/sizing_with_Rb_calculation.py
@@ -34,7 +34,7 @@ if __name__ == "__main__":
             peak_load_heating_array[i, j] = np.random.randint(monthly_load_heating_array[i, j], max_value_heating)
 
     # initiate borefield model
-    data = GroundData(100, 6, 3, 10, 0.2, 10, 12)  # ground data with an inaccurate guess of 100m for the depth of the borefield
+    data = GroundData(100, 6, 3, 10, 0.2, 10, 12, 2.4 * 10**6)  # ground data with an inaccurate guess of 100m for the depth of the borefield
     fluid_data = FluidData(0.2, 0.568, 998, 4180, 1e-3)
     pipe_data = PipeData(1, 0.015, 0.02, 0.4, 0.05, 0.075, 2)
 
