+++ paper.md
@@ -19,7 +19,7 @@ affiliations:
    index: 2
  - name: Solar-Institute Jülich, FH Aachen, Aachen, Germany
    index: 3
-date: 01 April 2022
+date: 05 August 2022
 bibliography: paper.bib
 
 ---
