### 2022-08-12 0388196
Changed location images for PyPi
