### 2022-08-16 5e65d5d
First version of the code for validation of the linear relation between imbalance and load. (Paper Emma)
