### 2022-08-17 d309558
Update README.md
Added PhD of Sharifi to the list of mentions of GHEtool
