### 2022-08-17 #4406 @crvernon
@wouterpeere - this is looking great. Just a few edits from my end that need to be addressed before we move forward:

- [ ] Please address the following invalid and missing DOIs:

> MISSING DOIs
>
> - 10.26868/25222708.2021.30180 may be a valid DOI for title: Validated combined first > and last year borefield sizing methodology
>
> INVALID DOIs
>
> - https://doi.org/10.1016/j.enconman.2014.11.009 is INVALID because of 'https://doi.org/' > prefix
> - https://doi.org/10.1016/j.energy.2017.01.097 is INVALID because of 'https://doi.org/' prefix

- [ ] A recommendation, not requirement.  You have tests in your package that could easily be executed via continuous integration (CI) via GitHub Actions.  Please consider setting CI up for your software so that on push and/or pull request your tests automatically execute.

Great progress so far. Keep up the good work!
