### 2022-08-17 #4406 @wouterpeere
Hi @crvernon the DOIs are addressed!

W.r.t. CI: I will look into this, but for now, I was (due to some administrator issues on my university laptop) not able to run the pytest-cov locally ...
