### 2022-08-17 #4406 @wouterpeere
@crvernon Right, I forgot about the paper changes.
The version is now v2.0.4, the DOI is also updated to this new version: 10.5281/zenodo.7004037
