### 2022-08-17 800670c
Fixed bug (issue #15)
