+++ GHEtool/Validation/sizing_method_comparison.py
@@ -63,7 +63,7 @@ if __name__ == "__main__":
     borefield.set_max_ground_temperature(16)   # maximum temperature
     borefield.set_min_ground_temperature(0)    # minimum temperature
 
-    # size according to l_2 method
+    # size according to L2 method
     start_L2 = time.time()
     for i in range(number_of_iterations):
         borefield.set_baseload_cooling(monthly_load_cooling_array[i])
@@ -90,5 +90,5 @@ if __name__ == "__main__":
     for i in range(number_of_iterations):
         difference_results[i] = results_L3[i] - results_L2[i]
 
-    print("The maximal difference between the sizing of L2 and L3 was:", np.round(np.max(difference_results), 3), "m or", np.round(np.max(difference_results) / results_L2[np.argmax(difference_results)] * 100, 3), "% w.r.t. the l_2 sizing.")
-    print("The mean difference between the sizing of L2 and L3 was:", np.round(np.mean(difference_results), 3), "m or", np.round(np.mean(difference_results) / np.mean(results_L2) * 100, 3), "% w.r.t. the l_2 sizing.")
+    print("The maximal difference between the sizing of L2 and L3 was:", np.round(np.max(difference_results), 3), "m or", np.round(np.max(difference_results) / results_L2[np.argmax(difference_results)] * 100, 3), "% w.r.t. the L2 sizing.")
+    print("The mean difference between the sizing of L2 and L3 was:", np.round(np.mean(difference_results), 3), "m or", np.round(np.mean(difference_results) / np.mean(results_L2) * 100, 3), "% w.r.t. the L2 sizing.")
