### 2022-08-17 #4406 @kyleniemeyer
Hi @wouterpeere, I made a few minor edits to the paper: https://github.com/wouterpeere/GHEtool/pull/16

Can you merge these? I will then accept your submission.
