### 2022-08-17 #4406 @wouterpeere
> Hi @wouterpeere, I made a few minor edits to the paper: [wouterpeere/GHEtool#16](https://github.com/wouterpeere/GHEtool/pull/16)
>
> Can you merge these? I will then accept your submission.

Hi @kyleniemeyer , I merged you're changes. Do I need to update the archive and version also?
