### 2022-08-17 #4406 @kyleniemeyer
Congratulations @wouterpeere on your article's publication in JOSS!5981))
